requirements.txt
click
pyyaml
requests
python-dotenv